Add Python 3.x compatibility
Added six and future modules to handle both 2.x and 3.x.
Switched from wxPython to Phoenix. Replaced deprecated wx functions.

+++ tests/big_import.py
@@ -1,5 +1,6 @@
 """Module which will require a lot of time to import"""
+from builtins import int
 
 for i in range( 2500 ):
-	long(i) ** i
+    int(i) ** i
